src/requirements.txt
beautifulsoup4==4.12.3
html2text==2024.2.26
requests==2.32.3
urllib3==2.2.2
lxml==5.2.2
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
transformers==4.43.3
accelerate==0.33.0
torch>=2.1.0
datasets==2.20.0
peft==0.12.0
trl==0.9.6
bitsandbytes==0.43.3
fastmcp==0.4.0
requests==2.32.3
websockets==12.0

